+++ CHANGELOG.md
@@ -18,6 +18,21 @@ to [Semantic Versioning](https://semver.org).
 
 ### Security
 
+## 600.0.0+242 - 2025-01-30
+
+### Added
+
+- Code completion
+  for most
+  of [the value providers](https://docs.spring.io/spring-boot/docs/current/reference/html/configuration-metadata.html#configuration-metadata.manual-hints.value-providers).
+- Auto popup value completion after key insertion.
+- Better setter navigate: from configuration property source to configuration file.
+
+### Changed
+
+- Performance improvement: It is no longer necessary to load the Spring configuration metadata each time the project is
+  opened.
+
 ## 242.2.0 - 2024-09-18
 
 ### Fixed